+++ somni-parser/src/ast.rs
@@ -1,32 +1,40 @@
-use crate::{
-    lexer::Token,
-    parser::{DefaultTypeSet, TypeSet},
-    Location,
-};
+use crate::{lexer::Token, parser::TypeSet, Location};
 
 #[derive(Debug)]
-pub struct Program {
-    pub items: Vec<Item>,
+pub struct Program<T>
+where
+    T: TypeSet,
+{
+    pub items: Vec<Item<T>>,
 }
 
 #[derive(Debug)]
-pub enum Item {
-    Function(Function),
+pub enum Item<T>
+where
+    T: TypeSet,
+{
+    Function(Function<T>),
     ExternFunction(ExternalFunction),
-    GlobalVariable(GlobalVariable),
+    GlobalVariable(GlobalVariable<T>),
 }
 
 #[derive(Debug)]
-pub struct GlobalVariable {
+pub struct GlobalVariable<T>
+where
+    T: TypeSet,
+{
     pub decl_token: Token,
     pub identifier: Token,
     pub colon: Token,
     pub type_token: TypeHint,
     pub equals_token: Token,
-    pub initializer: Expression,
+    pub initializer: Expression<T>,
     pub semicolon: Token,
 }
-impl GlobalVariable {
+impl<T> GlobalVariable<T>
+where
+    T: TypeSet,
+{
     pub fn location(&self) -> Location {
         let start = self.decl_token.location.start;
         let end = self.semicolon.location.end;
@@ -53,21 +61,40 @@ pub struct ExternalFunction {
 }
 
 #[derive(Debug)]
-pub struct Function {
+pub struct Function<T>
+where
+    T: TypeSet,
+{
     pub fn_token: Token,
     pub name: Token,
     pub opening_paren: Token,
     pub arguments: Vec<FunctionArgument>,
     pub closing_paren: Token,
     pub return_decl: Option<ReturnDecl>,
-    pub body: Body,
+    pub body: Body<T>,
 }
 
-#[derive(Debug, Clone)]
-pub struct Body {
+#[derive(Debug)]
+pub struct Body<T>
+where
+    T: TypeSet,
+{
     pub opening_brace: Token,
-    pub statements: Vec<Statement>,
+    pub statements: Vec<Statement<T>>,
     pub closing_brace: Token,
+}
+
+impl<T> Clone for Body<T>
+where
+    T: TypeSet,
+{
+    fn clone(&self) -> Self {
+        Self {
+            opening_brace: self.opening_brace.clone(),
+            statements: self.statements.clone(),
+            closing_brace: self.closing_brace.clone(),
+        }
+    }
 }
 
 #[derive(Debug)]
@@ -83,17 +110,39 @@ pub struct TypeHint {
     pub type_name: Token,
 }
 
-#[derive(Debug, Clone)]
-pub struct VariableDefinition {
+#[derive(Debug)]
+pub struct VariableDefinition<T>
+where
+    T: TypeSet,
+{
     pub decl_token: Token,
     pub identifier: Token,
     pub type_token: Option<TypeHint>,
     pub equals_token: Token,
-    pub initializer: Expression,
+    pub initializer: Expression<T>,
     pub semicolon: Token,
 }
 
-impl VariableDefinition {
+impl<T> Clone for VariableDefinition<T>
+where
+    T: TypeSet,
+{
+    fn clone(&self) -> Self {
+        Self {
+            decl_token: self.decl_token.clone(),
+            identifier: self.identifier.clone(),
+            type_token: self.type_token.clone(),
+            equals_token: self.equals_token.clone(),
+            initializer: self.initializer.clone(),
+            semicolon: self.semicolon.clone(),
+        }
+    }
+}
+
+impl<T> VariableDefinition<T>
+where
+    T: TypeSet,
+{
     pub fn location(&self) -> Location {
         let start = self.decl_token.location.start;
         let end = self.semicolon.location.end;
@@ -101,14 +150,33 @@ impl VariableDefinition {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct ReturnWithValue {
+#[derive(Debug)]
+pub struct ReturnWithValue<T>
+where
+    T: TypeSet,
+{
     pub return_token: Token,
-    pub expression: Expression,
+    pub expression: Expression<T>,
     pub semicolon: Token,
 }
 
-impl ReturnWithValue {
+impl<T> Clone for ReturnWithValue<T>
+where
+    T: TypeSet,
+{
+    fn clone(&self) -> Self {
+        Self {
+            return_token: self.return_token.clone(),
+            expression: self.expression.clone(),
+            semicolon: self.semicolon.clone(),
+        }
+    }
+}
+
+impl<T> ReturnWithValue<T>
+where
+    T: TypeSet,
+{
     pub fn location(&self) -> Location {
         let start = self.return_token.location.start;
         let end = self.semicolon.location.end;
@@ -130,18 +198,50 @@ impl EmptyReturn {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct If {
+#[derive(Debug)]
+pub struct If<T>
+where
+    T: TypeSet,
+{
     pub if_token: Token,
-    pub condition: Expression,
-    pub body: Body,
-    pub else_branch: Option<Else>,
+    pub condition: Expression<T>,
+    pub body: Body<T>,
+    pub else_branch: Option<Else<T>>,
 }
 
-#[derive(Debug, Clone)]
-pub struct Loop {
+impl<T> Clone for If<T>
+where
+    T: TypeSet,
+{
+    fn clone(&self) -> Self {
+        Self {
+            if_token: self.if_token.clone(),
+            condition: self.condition.clone(),
+            body: self.body.clone(),
+            else_branch: self.else_branch.clone(),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct Loop<T>
+where
+    T: TypeSet,
+{
     pub loop_token: Token,
-    pub body: Body,
+    pub body: Body<T>,
+}
+
+impl<T> Clone for Loop<T>
+where
+    T: TypeSet,
+{
+    fn clone(&self) -> Self {
+        Self {
+            loop_token: self.loop_token.clone(),
+            body: self.body.clone(),
+        }
+    }
 }
 
 #[derive(Debug, Clone)]
@@ -156,29 +256,71 @@ pub struct Continue {
     pub semicolon: Token,
 }
 
-#[derive(Debug, Clone)]
-pub enum Statement {
-    VariableDefinition(VariableDefinition),
-    Return(ReturnWithValue),
+#[derive(Debug)]
+pub enum Statement<T>
+where
+    T: TypeSet,
+{
+    VariableDefinition(VariableDefinition<T>),
+    Return(ReturnWithValue<T>),
     EmptyReturn(EmptyReturn),
-    If(If),
-    Loop(Loop),
+    If(If<T>),
+    Loop(Loop<T>),
     Break(Break),
     Continue(Continue),
     Expression {
-        expression: Expression,
+        expression: Expression<T>,
         semicolon: Token,
     },
 }
 
-#[derive(Debug, Clone)]
-pub struct Else {
-    pub else_token: Token,
-    pub else_body: Body,
+impl<T> Clone for Statement<T>
+where
+    T: TypeSet,
+{
+    fn clone(&self) -> Self {
+        match self {
+            Self::VariableDefinition(arg0) => Self::VariableDefinition(arg0.clone()),
+            Self::Return(arg0) => Self::Return(arg0.clone()),
+            Self::EmptyReturn(arg0) => Self::EmptyReturn(arg0.clone()),
+            Self::If(arg0) => Self::If(arg0.clone()),
+            Self::Loop(arg0) => Self::Loop(arg0.clone()),
+            Self::Break(arg0) => Self::Break(arg0.clone()),
+            Self::Continue(arg0) => Self::Continue(arg0.clone()),
+            Self::Expression {
+                expression,
+                semicolon,
+            } => Self::Expression {
+                expression: expression.clone(),
+                semicolon: semicolon.clone(),
+            },
+        }
+    }
 }
 
-#[derive(Debug, Clone)]
-pub enum Expression<T = DefaultTypeSet>
+#[derive(Debug)]
+pub struct Else<T>
+where
+    T: TypeSet,
+{
+    pub else_token: Token,
+    pub else_body: Body<T>,
+}
+
+impl<T> Clone for Else<T>
+where
+    T: TypeSet,
+{
+    fn clone(&self) -> Self {
+        Self {
+            else_token: self.else_token.clone(),
+            else_body: self.else_body.clone(),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub enum Expression<T>
 where
     T: TypeSet,
 {
@@ -200,6 +342,34 @@ where
         name: Token,
         arguments: Box<[Self]>,
     },
+}
+
+impl<T> Clone for Expression<T>
+where
+    T: TypeSet,
+{
+    fn clone(&self) -> Self {
+        match self {
+            Self::Variable { variable } => Self::Variable {
+                variable: variable.clone(),
+            },
+            Self::Literal { value } => Self::Literal {
+                value: value.clone(),
+            },
+            Self::UnaryOperator { name, operand } => Self::UnaryOperator {
+                name: name.clone(),
+                operand: operand.clone(),
+            },
+            Self::BinaryOperator { name, operands } => Self::BinaryOperator {
+                name: name.clone(),
+                operands: operands.clone(),
+            },
+            Self::FunctionCall { name, arguments } => Self::FunctionCall {
+                name: name.clone(),
+                arguments: arguments.clone(),
+            },
+        }
+    }
 }
 impl<T> Expression<T>
 where
@@ -248,8 +418,8 @@ where
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct Literal<T = DefaultTypeSet>
+#[derive(Debug)]
+pub struct Literal<T>
 where
     T: TypeSet,
 {
@@ -257,8 +427,20 @@ where
     pub location: Location,
 }
 
-#[derive(Debug, Clone)]
-pub enum LiteralValue<T = DefaultTypeSet>
+impl<T> Clone for Literal<T>
+where
+    T: TypeSet,
+{
+    fn clone(&self) -> Self {
+        Self {
+            value: self.value.clone(),
+            location: self.location.clone(),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub enum LiteralValue<T>
 where
     T: TypeSet,
 {
@@ -266,4 +448,18 @@ where
     Float(T::Float),
     String(String),
     Boolean(bool),
+}
+
+impl<T> Clone for LiteralValue<T>
+where
+    T: TypeSet,
+{
+    fn clone(&self) -> Self {
+        match self {
+            Self::Integer(arg0) => Self::Integer(arg0.clone()),
+            Self::Float(arg0) => Self::Float(arg0.clone()),
+            Self::String(arg0) => Self::String(arg0.clone()),
+            Self::Boolean(arg0) => Self::Boolean(arg0.clone()),
+        }
+    }
 }