+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod codegen;
 pub mod error;
 pub mod ir;
+pub mod string_interner;
 #[cfg(test)]
 pub mod test;
 pub mod transform_ir;