+++ src/transform_ir.rs
@@ -1,11 +1,11 @@
 use std::collections::{HashMap, HashSet};
 
 use indexmap::IndexMap;
-use somni_expr::string_interner::{self, StringIndex};
 
 use crate::{
     error::CompileError,
     ir::{self, BlockIndex, VariableIndex},
+    string_interner::{self, StringIndex},
     variable_tracker::{LocalVariableIndex, ScopeData},
 };
 use somni_parser::Location;