+++ src/types.rs
@@ -1,4 +1,7 @@
-use somni_expr::{string_interner::StringIndex, Type, TypeSet};
+use somni_expr::{OperatorError, Type, TypeSet};
+use somni_parser::parser::TypeSet as ParserTypeSet;
+
+use crate::string_interner::StringIndex;
 
 #[doc(hidden)]
 pub trait MemoryRepr: Sized + Copy + PartialEq {
@@ -58,31 +61,137 @@ mem_repr!(f32);
 mem_repr!(f64);
 
 pub(crate) trait TypeExt {
-    /// Returns the size of the type in bytes.
-    fn size_of<T>(&self) -> usize
-    where
-        T: TypeSet,
-        T::Integer: MemoryRepr,
-        T::SignedInteger: MemoryRepr,
-        T::Float: MemoryRepr;
+    /// Returns the size of the type in bytes in the VM.
+    fn vm_size_of(&self) -> usize;
 }
 
 impl TypeExt for somni_expr::Type {
-    /// Returns the size of the type in bytes.
-    fn size_of<T>(&self) -> usize
-    where
-        T: TypeSet,
-        T::Integer: MemoryRepr,
-        T::SignedInteger: MemoryRepr,
-        T::Float: MemoryRepr,
-    {
+    fn vm_size_of(&self) -> usize {
         match self {
             Type::Void => <() as MemoryRepr>::BYTES,
-            Type::Int | Type::MaybeSignedInt => <T::Integer as MemoryRepr>::BYTES,
-            Type::SignedInt => <T::SignedInteger as MemoryRepr>::BYTES,
-            Type::Float => <T::Float as MemoryRepr>::BYTES,
+            Type::Int | Type::MaybeSignedInt => {
+                <<VmTypeSet as TypeSet>::Integer as MemoryRepr>::BYTES
+            }
+            Type::SignedInt => <<VmTypeSet as TypeSet>::SignedInteger as MemoryRepr>::BYTES,
+            Type::Float => <<VmTypeSet as TypeSet>::Float as MemoryRepr>::BYTES,
             Type::Bool => <bool as MemoryRepr>::BYTES,
-            Type::String => <StringIndex as MemoryRepr>::BYTES,
+            Type::String => <<VmTypeSet as TypeSet>::String as MemoryRepr>::BYTES,
         }
+    }
+}
+
+/// Represents any value inside the VM.
+#[doc(hidden)]
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum TypedValue {
+    /// Represents no value.
+    Void,
+    /// Represents an integer that may be signed or unsigned.
+    MaybeSignedInt(u64),
+    /// Represents an unsigned integer.
+    Int(u64),
+    /// Represents a signed integer.
+    SignedInt(i64),
+    /// Represents a floating-point.
+    Float(f64),
+    /// Represents a boolean.
+    Bool(bool),
+    /// Represents a string.
+    String(StringIndex),
+}
+
+impl From<TypedValue> for somni_expr::TypedValue<VmTypeSet> {
+    fn from(value: TypedValue) -> Self {
+        match value {
+            TypedValue::Void => somni_expr::TypedValue::Void,
+            TypedValue::MaybeSignedInt(inner) => somni_expr::TypedValue::MaybeSignedInt(inner),
+            TypedValue::Int(inner) => somni_expr::TypedValue::Int(inner),
+            TypedValue::SignedInt(inner) => somni_expr::TypedValue::SignedInt(inner),
+            TypedValue::Float(inner) => somni_expr::TypedValue::Float(inner),
+            TypedValue::Bool(inner) => somni_expr::TypedValue::Bool(inner),
+            TypedValue::String(inner) => somni_expr::TypedValue::String(inner),
+        }
+    }
+}
+impl From<somni_expr::TypedValue<VmTypeSet>> for TypedValue {
+    fn from(value: somni_expr::TypedValue<VmTypeSet>) -> Self {
+        match value {
+            somni_expr::TypedValue::Void => TypedValue::Void,
+            somni_expr::TypedValue::MaybeSignedInt(inner) => TypedValue::MaybeSignedInt(inner),
+            somni_expr::TypedValue::Int(inner) => TypedValue::Int(inner),
+            somni_expr::TypedValue::SignedInt(inner) => TypedValue::SignedInt(inner),
+            somni_expr::TypedValue::Float(inner) => TypedValue::Float(inner),
+            somni_expr::TypedValue::Bool(inner) => TypedValue::Bool(inner),
+            somni_expr::TypedValue::String(inner) => TypedValue::String(inner),
+        }
+    }
+}
+
+impl TypedValue {
+    pub fn type_of(&self) -> somni_expr::Type {
+        match self {
+            TypedValue::Void => Type::Void,
+            TypedValue::MaybeSignedInt(_) => Type::MaybeSignedInt,
+            TypedValue::Int(_) => Type::Int,
+            TypedValue::SignedInt(_) => Type::SignedInt,
+            TypedValue::Float(_) => Type::Float,
+            TypedValue::Bool(_) => Type::Bool,
+            TypedValue::String(_) => Type::String,
+        }
+    }
+
+    pub(crate) fn write(&self, to: &mut [u8]) {
+        match self {
+            TypedValue::Void => ().write(to),
+            TypedValue::Int(value) | TypedValue::MaybeSignedInt(value) => value.write(to),
+            TypedValue::SignedInt(value) => value.write(to),
+            TypedValue::Float(value) => value.write(to),
+            TypedValue::Bool(value) => value.write(to),
+            TypedValue::String(value) => value.write(to),
+        }
+    }
+
+    pub(crate) fn from_typed_bytes(ty: Type, value: &[u8]) -> TypedValue {
+        match ty {
+            Type::Void => Self::Void,
+            Type::Int => Self::Int(<_>::from_bytes(value)),
+            Type::MaybeSignedInt => Self::MaybeSignedInt(<_>::from_bytes(value)),
+            Type::SignedInt => Self::SignedInt(<_>::from_bytes(value)),
+            Type::Float => Self::Float(<_>::from_bytes(value)),
+            Type::Bool => Self::Bool(<_>::from_bytes(value)),
+            Type::String => Self::String(<_>::from_bytes(value)),
+        }
+    }
+}
+
+#[derive(Debug, Default)]
+pub struct VmTypeSet;
+
+impl ParserTypeSet for VmTypeSet {
+    type Integer = u64;
+    type Float = f64;
+}
+
+impl TypeSet for VmTypeSet {
+    type Parser = Self;
+
+    type Integer = u64;
+    type SignedInteger = i64;
+    type Float = f64;
+    type String = StringIndex;
+
+    fn to_signed(v: Self::Integer) -> Result<Self::SignedInteger, OperatorError> {
+        i64::try_from(v).map_err(|_| OperatorError::RuntimeError)
+    }
+
+    fn load_string<'s>(
+        _ctx: &'s dyn somni_expr::ExprContext<Self>,
+        _str: &'s Self::String,
+    ) -> &'s str {
+        todo!()
+    }
+
+    fn store_string(_ctx: &mut dyn somni_expr::ExprContext<Self>, _str: &str) -> Self::String {
+        todo!()
     }
 }