+++ src/variable_tracker.rs
@@ -1,7 +1,6 @@
 use indexmap::{IndexMap, IndexSet};
 
-use crate::ir::Variable;
-use somni_expr::string_interner::StringIndex;
+use crate::{ir::Variable, string_interner::StringIndex};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct RestorePoint(usize);
@@ -195,7 +194,7 @@ impl VariableTracker {
 mod tests {
     use super::*;
 
-    use somni_expr::string_interner::StringInterner;
+    use crate::string_interner::StringInterner;
 
     #[test]
     fn test_variable_tracker() {