+++ somni-expr/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Unreleased
 
 - The TypeSet trait has been reworked for better ergonomics. The parser's TypeSet is now an associated type. Additional associated types and trait bounds have been added.
-- The `MemoryRepr` trait has been removed
+- The `MemoryRepr` trait has been removed.
+- TypedValue's `String` is now also generic.
+- The string interner has been removed. The expression engine uses String as storage by default.
 
 # [0.1.0] - 2025-07-24