+++ somni-parser/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Hidden the tokenizer function
 - Unified lexer/parser errors
 - Parser no longer needs the whole source to be tokenized upfront
+- Added TypeSet type parameter to the whole AST and the `parse` function
 
 # [0.1.0] - 2025-07-24